Add non-formatting versions of the logging functions
During creation of the telemetry buffer, the message has to be
duplicated in memory. When formatting, the format message has to be
duplicated again. Especially in situations, where no formatting is
necessary, but the input string could contain formatting characters,
this is very wasteful and can lead to stack overflows.

By using the non-formatting versions, one of the copies in memory is
avoided.